requirements.txt
# Packages needed to test and run the app.
hashin==0.3 \
    --hash=sha256:e86f8b9104df83bda4cf8d000bf21b7fde1ebf3962e67b03fea4c102970d2f95 \
    --hash=sha256:8015468f50b97cfd4fc8d81fdd99966f1959d8990cee27e24aa79976a2f4ebc1 \
    --hash=sha256:f8c666b790da4327622ac492fdbcf6caa1b46a0d85fae90eada5213372fafe97
argparse==1.4.0 \
    --hash=sha256:c31647edb69fd3d465a847ea3157d37bed1f95f19760b11a47aa91c04b666314 \
    --hash=sha256:62b089a55be1d8949cd2bc7e0df0bddb9e028faefc8c32038cc84862aefdd6e4
Django==1.9.1 \
    --hash=sha256:9f7ca04c6dbcf08b794f2ea5283c60156a37ebf2b8316d1027f594f34ff61101 \
    --hash=sha256:a29aac46a686cade6da87ce7e7287d5d53cddabc41d777c6230a583c36244a18
dj-database-url==0.3.0 \
    --hash=sha256:f2e273ed34acbb560962d5cf12917936d8df02297df09bd3089b8546d4584138 \
    --hash=sha256:ca01768fdecde134301f3170743226f60edff5c3935f12437378ebd911506353
psycopg2==2.6.1 \
    --hash=sha256:6acf9abbbe757ef75dc2ecd9d91ba749547941abaffbe69ff2086a9e37d4904c
python-decouple==3.0 \
    --hash=sha256:99834c9ff7ce5c5b2f4a18bc0880753e54ea3aaabc3cfb16961a92d5046665df
gunicorn==19.3.0 \
    --hash=sha256:a5179cde922d2b4e045ee5b11e87323ee77d363ae40294fc8252f25d6a0eaf06 \
    --hash=sha256:8bc835082882ad9a012cd790c460011e4d96bf3512d98a04d3dabbe45393a089
python-ldap==2.4.25 \
    --hash=sha256:62d00dbc86f3f9b21beacd9b826e8f9895f900637a60a6d4e7ab59a1cdc64e56 \
    --hash=sha256:28f413b8a09dbe554958e4aa957c00afc2159011ae3a01b273b718a65fed4d06 \
    --hash=sha256:5837c1cd93180e441291654c10572558da9fef244b5d5db2799c84551820f283 \
    --hash=sha256:043314ef7103cc81ae9d86584a8bf022d892d2297eb83d9c5c356d16bdf59f0c \
    --hash=sha256:4c158905fa0ff71b666d360eda0cd44c304bfe8d8b5e45a416b3cd4c4282f3d5
django-session-csrf==0.6 \
    --hash=sha256:009d574590aeebff230bda3d8c1cacd9d48b204cc1a784a5d3be3638727f8d16 \
    --hash=sha256:ccfddfbf947a51c8b095a5e63bed43005a277a5581de92ed373eb06b75ceec77
django-jsonview==0.5.0 \
    --hash=sha256:d1544e97dd5e625a77f9539002bcdec3e12c673bdadd491271782ef29016470a \
    --hash=sha256:58cc5d5d992ae3fe36084c2eb103e02ec5fbb5470736df17f96b5c3bdc7c8138
